+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Fixed:
 - Fixed mechanical piston pushing
 - Fixed placement of ammunition in autocannon ammo container screen
 - Fixed crash with CBC: Advanced Technologies
+- Fixed clipboard not correctly copying Fixed Cannon Mount angles
 - [1.21.1] Fixed incorrect block data saving leading to stack overflow crashes
 - [Aeronautics] Fixed cannon blocks and cannon contraptions not attaching to ship on assembly
 - [Aeronautics] Fixed Cannon Loader, Cannon Drill, and Cannon Builder not attaching to ship on assembly